chore: reorganise all Api finalisation steps in a class [DHIS2-15360]
